backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
groq>=0.4.1
PyGithub>=1.59.1
gitpython>=3.1.40
pyyaml>=6.0.1
python-multipart>=0.0.6
structlog>=23.2.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
